% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdsr.R
\name{mdsr_exercise_new}
\alias{mdsr_exercise_new}
\alias{mdsr_exercise_ls}
\alias{mdsr_exercise_all}
\alias{knit_mdsr_exercises}
\alias{knit_mdsr_exercises_filter}
\alias{knit_mdsr_exercises_print}
\alias{knit_mdsr_exercises_onlineonly}
\alias{knit_mdsr_exercises_solutions}
\alias{include_mdsr_img}
\title{List, filter, and show MDSR2e exercises}
\usage{
mdsr_exercise_new(directory = here::here("inst", "Exercises"), learnr = FALSE)

mdsr_exercise_ls(...)

mdsr_exercise_all()

knit_mdsr_exercises(x, show_answers = FALSE)

knit_mdsr_exercises_filter(..., show_answers = FALSE)

knit_mdsr_exercises_print(..., show_answers = FALSE)

knit_mdsr_exercises_onlineonly(..., show_answers = FALSE)

knit_mdsr_exercises_solutions(...)

include_mdsr_img(img)
}
\arguments{
\item{directory}{Path to the directory where the files go}

\item{learnr}{Logical flag (default: FALSE) whether to use
a learnr compatible document for the exercise.}

\item{...}{arguments passed to \code{\link[dplyr]{filter}}}

\item{x}{data frame of exercises}

\item{show_answers}{Should the answers be shown?}

\item{url}{URL of an image from the Internet}
}
\description{
List, filter, and show MDSR2e exercises
}
\examples{
\dontrun{
mdsr_exercise_new()
}
mdsr_exercise_ls()
include_mdsr_img("hickey-denzel-1.png")
include_mdsr_img("dplyr-arrange.png")

}
\seealso{
\code{\link[etude]{new_etude}}

\code{\link[etude]{etude_dir}}, \code{\link[dplyr]{filter}}
}
