% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macros.R
\name{pkg_macro}
\alias{pkg_macro}
\alias{df_macro}
\alias{var_macro}
\alias{val_macro}
\alias{func_macro}
\title{Macros for typesetting in the MDSR2 exercises}
\usage{
pkg_macro(word)

df_macro(word)

var_macro(word)

val_macro(word)

func_macro(word)
}
\arguments{
\item{word}{the word you want to use the macro on}
}
\description{
Provides inline hooks for packages, data frames, and ....
}
