% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{render_html}
\alias{render_html}
\title{Capture HTML output to clipboard}
\usage{
render_html(file)
}
\arguments{
\item{file}{path to an exercise file}
}
\description{
Copy the HTML output from a particular question to the clipboard
}
\examples{
file <- mdsr_exercise_ls()$file[4]
render_html(file)
}
