% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{write_moodle}
\alias{write_moodle}
\title{R/exams integration}
\usage{
write_moodle(pattern = "*.Rmd", quiz_name = "mdsr_quiz", dir = ".", ...)
}
\arguments{
\item{pattern}{Regular expression passed to \code{\link[fs]{dir_ls}}}

\item{quiz_name}{Character giving the name of the quiz}

\item{dir}{Directory where you want the quiz file to go}

\item{...}{arguments passed to \code{\link[exams]{exams2moodle}}}
}
\description{
R/exams integration
}
\examples{
\dontrun{
write_moodle(pattern = "ethics", "mdsr_ethics", dir = tempdir())
}
}
